refactor(indexing): one walker for "which .rs files belong to a project"
The walk lived inside `unified_parallel::collect_rust_files`, reachable only by
the indexer. A daemon that keeps a rust-analyzer context loaded across sessions
needs the same answer for a different question — "did anything in this project
change while it sat idle?" — and two independent walkers would mean the two
sides disagree about what a project file is, which surfaces as a context that is
either refreshed for no reason or not refreshed when it should be.

No behaviour change: the skip list and the traversal are the same ones, moved
into `indexing::traversal` and now shared. `collect_rust_files` keeps its
`stats` bookkeeping and delegates the walking.
